Fixes race condition.
Previously the loading branch component wasn't doing anything, but appsync was fast enough that it didn't matter to the application.  With more complex requests, the api is taking longer to respond revealing the problem.  This commit sets the loading prop to activate the loading branch in qewl and stop the race.