Upgrades Qewl onSubmit to return apollo promise (#18)
* Upgrades Qewl onSubmit to return apollo promise

Previously qewl called apollo-react's mutation function, but returned undefined.  Now it returns the promise from apollo, allowing us to do stuff like await the mutation before kicking the user somewhere else.

* 1.0.2

* Fixes race condition.

Previously the loading branch component wasn't doing anything, but appsync was fast enough that it didn't matter to the application.  With more complex requests, the api is taking longer to respond revealing the problem.  This commit sets the loading prop to activate the loading branch in qewl and stop the race.